Tests/LibWeb/Text/expected/navigation/ui-process-session-history-dump.txt

after-navigation: location=?second
after-navigation: ui-current=?second
after-navigation: ui-current-entry=?second*
after-navigation: ui-webcontent-known-current-entry=?second*
after-navigation: ui-forward-button=false
